requirements.txt
Flask-SQLAlchemy
flask_cors~=3.0.10
configparser~=5.0.2
logging
sqlalchemy~=2.0.31
pymysql
contextlib
tenacity
dbutils
mongoengine
requests~=2.31.0
json
config_control
flask~=3.0.3
pyngrok~=7.1.6
stripe~=10.4.0
python-dotenv~=1.0.1
pymongo~=4.8.0
apscheduler~=3.9.0
tika==1.24
Werkzeug==2.1.2
requests~=2.31.0